+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/common/Alignment.kt
@@ -1,0 +1,3 @@
+package com.agileburo.anytype.core_ui.common
+
+enum class Alignment { START, CENTER, END }