+++ presentation/src/main/java/com/agileburo/anytype/presentation/page/Editor.kt
@@ -1,5 +1,6 @@
 package com.agileburo.anytype.presentation.page
 
+import com.agileburo.anytype.core_ui.features.page.BlockView
 import com.agileburo.anytype.domain.editor.Editor.Focus
 import com.agileburo.anytype.presentation.page.editor.Proxy
 import com.agileburo.anytype.presentation.page.editor.Store
@@ -8,6 +9,7 @@ import com.agileburo.anytype.presentation.page.selection.SelectionStateHolder
 interface Editor {
 
     class Storage {
+        val views: Store<List<BlockView>> = Store.Screen()
         val focus: Store<Focus> = Store.Focus()
         val details: Store.Details = Store.Details()
     }