+++ presentation/src/main/java/com/agileburo/anytype/presentation/page/editor/Store.kt
@@ -1,10 +1,12 @@
 package com.agileburo.anytype.presentation.page.editor
 
+import com.agileburo.anytype.core_ui.features.page.BlockView
 import com.agileburo.anytype.domain.block.model.Block
 import com.agileburo.anytype.domain.common.Id
 import com.agileburo.anytype.domain.editor.Editor
 import kotlinx.coroutines.channels.ConflatedBroadcastChannel
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asFlow
 import timber.log.Timber
 
@@ -31,6 +33,16 @@ interface Store<T> {
 
     fun cancel()
 
+    open class State<T>(initial : T) : Store<T> {
+
+        private val state = MutableStateFlow(initial)
+
+        override fun stream(): Flow<T> = state
+        override fun current(): T = state.value
+        override suspend fun update(t: T) { state.value = t }
+        override fun cancel() {}
+    }
+
     open class Conflated<T>(default: T) : Store<T> {
 
         private val channel = ConflatedBroadcastChannel(default)
@@ -49,6 +61,7 @@ interface Store<T> {
         }
     }
     class Context : Conflated<String>("")
+    class Screen : State<List<BlockView>>(emptyList())
 
     class Details : Conflated<Block.Details>(Block.Details()) {
         suspend fun add(target: Id, fields: Block.Fields) {